Fix problem when transitioning from ultra wide display
When using a HP laptop with Intel UHD integrated graphics
and connecting to an ultra wide display with the lid down
will cause the ultrawide resolution to be applied to the next
display that the computer is connected to even if the display
mode is not a natively supported resolution for the display.

This behaviour was discovered by starting a HP EliteBook with
the lid down while connected to a 3440x1440 display (Dell or HP).
After logging in the laptop was turned off.

Then laptop was then started again with the lid down but connected
to two daisy-chaned displays with 2560X1440 resolution (a "HP E27m G4"
and "HP E27q G4"). After logging in the resolution on the first screen
was set to 3440x1440 by Windows 10/11 which is not a supported native
resolution for the display (HP E27m G4). In fact, using the display
configuration tool on Windows 10/11 would report that this was not
the optimal resolution for the display.

The app retrieve a display mode that corresponds to the current screen
settings. The fix here involves checking if the selected display mode
for the screen is natively supported by the screen itself. If the mode
is not supported, I assign the display mode the lowest possible score (0)
so that the app completely disregards it. Instead, the app only utilizes
display modes that are supported by the screen.

For example, if I have an ultrawide resolution from a previously
connected screen configured on a screen that doesn't support that resolution,
the app ignores it and utilizes the optimal resolution reported by the
 operating system based on the display modes that the screen can actually handle.